requirements.txt
Package           Version
----------------- -------
annotated-types   0.5.0
anyio             3.7.1
click             8.1.6
databases         0.7.0
exceptiongroup    1.1.2
fastapi           0.100.0
greenlet          2.0.2
h11               0.14.0
idna              3.4
pip               23.2.1
psycopg2-binary   2.9.6
pydantic          2.1.0
pydantic_core     2.4.0
python-dotenv     1.0.0
pytz              2023.3
setuptools        56.0.0
sniffio           1.3.0
SQLAlchemy        1.4.49
starlette         0.27.0
typing_extensions 4.7.1
uvicorn           0.23.1
asyncpg           0.28.0
